requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.3
python-multipart==0.0.9
qdrant-client==1.7.3
minio==7.2.7
numpy==1.26.4
pypdf==4.3.1
python-docx==1.1.2
# Optional for real embeddings (uncomment to use)
# sentence-transformers==3.0.1

